UI 개선
+ 일부 버그 수정
+ 크리티컬 / 회피 작용 이펙트 추가
+ 콤보 획득 이펙트 추가
+ HP BAR 추가
+ EXP BAR 추가
+ 일부 UI 변경

+++ Running_Animal/Assets/Script/Character.cs
@@ -23,7 +23,7 @@ public class Character
     public int LV_JUMP_COUNT;
     public float DEF;
     public int LV_DEF;
-    public float LUK;
+    public int LUK;
     public int LV_LUK;
     public DataManager.Active_Skil ACTIVE;
 
@@ -34,7 +34,7 @@ public class Character
                      float _DOWN_POWER, int _LV_DOWN_POWER,
                      float _JUMP_COUNT, int _LV_JUMP_COUNT,
                      float _DEF, int _LV_DEF,
-                     float _LUK, int _LV_LUK, DataManager.Active_Skil _ACTIVE)
+                     int _LUK, int _LV_LUK, DataManager.Active_Skil _ACTIVE)
     {
         LV = _LV;
         STAT_POINT = _STAT_POINT;
